Add `_txlock=exclusive` to database DSN
Fixes #26

When under a high write load, the database could deadlock and cause
the entire system to become unresponsive.

The default transaction locking mode is DEFERRED, meaning that the
transaction is only started after the first command, which also
determines the transaction type. Since most voting transactions
start with a read command, this would set the transaction mode
accordingly. The subsequent write transactions would then trigger
the transaction mode to be changed, which caused a deadlock as
multiple concurrent transactions were hitting this point at once.

To work around this, the transaction locking mode is set to
EXCLUSIVE, meaning that only one transaction can be started at a
time instead of the potentially multiple "read-only" of DEFERRED
mode.